# openapi.yaml
...
openapi: 3.1.0
info:
  title: Lentera — LMS Wakaf Salman API
  version: 1.0.0
  description: |-
    Kontrak berdasarkan implementasi Flask di repository. Gunakan Bearer JWT dari POST /api/auth/login. Logout tercatat melalui POST /api/auth/logout; token baru yang terkait sesi login tersebut ditolak setelah logout. Role dan status akun dibaca ulang dari database pada setiap request.

    Path learning memakai /api/course (singular); assessment memakai /api/courses (plural). ID berasal dari database, bukan ID contoh/frontend. Daftar katalog/user/enrollment tidak memiliki pagination; hanya activity mendukung pagination.

    Payload JSON memakai Content-Type: application/json. Field request yang tidak dikenal ditolak kecuali login/register yang mengabaikan field tambahan. Input teks umumnya di-trim. Global request limit 3 MiB; endpoint profil memakai batas 2 MiB + 64 KiB. Error aplikasi umumnya {message}; error framework/proxy (termasuk 413 di luar auth) dapat berupa HTML. Timestamp identitas/learning dapat tanpa timezone, sedangkan timestamp assessment/discussion memakai UTC dengan Z.

    Endpoint Development bersifat opsional dan selalu tidak tersedia pada production. Spesifikasi ini tidak menambahkan route HTTP baru atau Swagger UI.
servers:
- url: http://localhost:5000
  description: Flask development lokal
- url: http://localhost:8000
  description: Docker API lokal; sesuaikan API_PORT
- url: https://{hostname}
...
  variables:
    hostname:
      default: api.example.com
...
security:
- bearerAuth: []
paths:
...
  /api/admin/courses/{course_id}/test:
    parameters:
    - {name: course_id, in: path, required: true, schema: {type: integer, minimum: 1}}
    get:
      tags: [Assessment]
      operationId: getAdminTest
      summary: Soal, kunci, dan pengaturan course untuk admin
      x-required-role: admin
      responses:
        '200': {description: Course, passing_grade, maksimal_attempt, masa_tunggu_test_hari, questions}
        '403': {$ref: '#/components/responses/Error403'}
    patch:
      tags: [Assessment]
      operationId: updateAdminTest
      summary: Ubah pengaturan ujian course
      x-required-role: admin
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                passing_grade: {type: integer, minimum: 0, maximum: 100}
                maksimal_attempt: {type: integer, minimum: 1, maximum: 65535}
                masa_tunggu_test_hari: {type: integer, minimum: 0, maximum: 65535}
      responses:
        '200': {description: Pengaturan tersimpan}
        '400': {$ref: '#/components/responses/Error400'}
        '403': {$ref: '#/components/responses/Error403'}
...
  /api/courses/{course_id}/test:
    get:
      tags:
      - Assessment
      summary: Soal dan metadata final test
      operationId: get_api_courses_course_id_test
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestDetail'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
      description: test_id sama dengan course_id. Memerlukan course aktif dan enrollment. GET tidak membuat attempt;
        discussion test tetap dapat diakses sebelum materi selesai.
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
  /api/courses/{course_id}/questions/{question_id}:
    get:
      tags:
      - Assessment
      summary: Detail soal aktif
      operationId: get_api_courses_course_id_questions_question_id
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  question:
                    $ref: '#/components/schemas/Question'
                required:
                - question
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: question_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
  /api/courses/{course_id}/test-attempts:
    post:
      tags:
      - Assessment
      summary: Mulai atau lanjutkan attempt
      operationId: post_api_courses_course_id_test_attempts
      responses:
        '201':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt:
                    $ref: '#/components/schemas/Attempt'
                  questions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Question'
                required:
                - attempt
                - questions
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '200':
          description: Attempt aktif dilanjutkan, tanpa menambah jumlah attempt.
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt:
                    $ref: '#/components/schemas/Attempt'
                  questions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Question'
                required:
                - attempt
                - questions
      description: Body {} wajib. 200 melanjutkan attempt aktif, 201 membuat baru. Attempt baru memerlukan semua
        materi aktif selesai, belum lulus, belum mencapai maksimal attempt, dan cooldown telah berakhir. Snapshot
        soal/kunci/passing score dibekukan saat mulai.
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyObject'
  /api/test-attempts/{attempt_id}:
    get:
      tags:
      - Assessment
      summary: Detail attempt sendiri
      operationId: get_api_test_attempts_attempt_id
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt:
                    $ref: '#/components/schemas/Attempt'
                required:
                - attempt
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
  /api/test-attempts/{attempt_id}/answers:
    post:
      tags:
      - Assessment
      summary: Simpan atau perbarui jawaban
      operationId: post_api_test_attempts_attempt_id_answers
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt:
                    $ref: '#/components/schemas/Attempt'
                required:
                - attempt
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
      description: Hanya attempt sendiri yang in_progress. Validasi bulk atomik; soal duplikat ditolak.
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAnswers'
  /api/test-attempts/{attempt_id}/submit:
    post:
      tags:
      - Assessment
      summary: Submit dan hitung hasil ujian
      operationId: post_api_test_attempts_attempt_id_submit
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
      description: Body {} wajib jika memakai jawaban tersimpan. Hanya pemilik attempt aktif; submit kedua ditolak
        409. Kelulusan memperbarui enrollment, kegagalan dapat mengaktifkan cooldown.
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswers'
  /api/test-attempts/{attempt_id}/result:
    get:
      tags:
      - Assessment
      summary: Hasil attempt sendiri
      operationId: get_api_test_attempts_attempt_id_result
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
      description: 409 jika attempt belum completed.
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT HS256 dari login; Authorization: Bearer <token>. Berlaku 8 jam.'
  responses:
    Error400:
      description: Payload atau parameter tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error401:
      description: Token hilang/invalid/kedaluwarsa, akun tidak ditemukan, atau kredensial login salah.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: Akun nonaktif, role tidak sesuai, bukan pemilik, atau prasyarat akses belum terpenuhi.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: Data tidak ditemukan atau tidak dapat dilihat oleh pengguna.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: Konflik data, relasi, konfigurasi test, atau status operasi.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Data tidak valid
      required:
      - message
...
    Question:
      type: object
      properties:
        question_id:
          type: integer
          format: int64
          minimum: 1
        number:
          type: integer
        pertanyaan:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              option_id:
                type: integer
                format: int64
                minimum: 1
              text:
                type: string
            required:
            - option_id
            - text
      required:
      - question_id
      - number
      - pertanyaan
      - options
      description: Kunci jawaban tidak pernah disertakan.
    Answer:
      type: object
      properties:
        question_id:
          type: integer
          format: int64
          minimum: 1
        selected_option_id:
          type: integer
          format: int64
          minimum: 1
      required:
      - question_id
      - selected_option_id
      additionalProperties: false
    Attempt:
      type: object
      properties:
        attempt_id:
          type: integer
          format: int64
          minimum: 1
        course_id:
          type: integer
          format: int64
          minimum: 1
        test_id:
          type: integer
          format: int64
          minimum: 1
        attempt_number:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
          - in_progress
          - completed
        started_at:
          anyOf:
          - type: string
            format: date-time
            example: '2026-10-06T03:30:00Z'
          - type: 'null'
        completed_at:
          anyOf:
          - type: string
            format: date-time
            example: '2026-10-06T03:30:00Z'
          - type: 'null'
        answers:
          type: array
          items:
            $ref: '#/components/schemas/Answer'
      required:
      - attempt_id
      - course_id
      - test_id
      - attempt_number
      - status
      - started_at
      - completed_at
      - answers
    TestResult:
      type: object
      properties:
        attempt_id:
          type: integer
          format: int64
          minimum: 1
        course_id:
          type: integer
          format: int64
          minimum: 1
        test_id:
          type: integer
          format: int64
          minimum: 1
        total_questions:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        answered_questions:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        correct_answers:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        wrong_answers:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
        score:
          type: number
        passing_score:
          anyOf:
          - type: integer
          - type: 'null'
        passed:
          type: boolean
        completed_at:
          anyOf:
          - type: string
            format: date-time
            example: '2026-10-06T03:30:00Z'
          - type: 'null'
      required:
      - attempt_id
      - course_id
      - test_id
      - total_questions
      - answered_questions
      - correct_answers
      - wrong_answers
      - score
      - passing_score
      - passed
      - completed_at
      description: Untuk record legacy tanpa snapshot, total/answered/correct/wrong/passing_score bernilai null.
        Soal tidak dijawab dihitung salah. Score disimpan dua desimal; kelulusan dihitung sebelum pembulatan.
    TestDetail:
      type: object
      properties:
        course_id:
          type: integer
          format: int64
          minimum: 1
        test_id:
          type: integer
          format: int64
          minimum: 1
        passing_score:
          type: integer
        max_attempts:
          type: integer
        attempts_used:
          type: integer
          minimum: 0
        materials_completed:
          type: boolean
        test_access_status:
          type: string
          enum:
          - aktif
          - nonaktif
          - lulus
        retry_at:
          anyOf:
          - type: string
            format: date-time
            example: '2026-10-06T03:30:00Z'
          - type: 'null'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        active_attempt:
          anyOf:
          - $ref: '#/components/schemas/Attempt'
          - type: 'null'
        latest_result:
          anyOf:
          - $ref: '#/components/schemas/TestResult'
          - type: 'null'
      required:
      - course_id
      - test_id
      - passing_score
      - max_attempts
      - attempts_used
      - materials_completed
      - test_access_status
      - retry_at
      - questions
      - active_attempt
      - latest_result
    EmptyObject:
      type: object
      properties: {}
      additionalProperties: false
      example: {}
    SaveAnswers:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: '#/components/schemas/Answer'
          minItems: 1
      required:
      - answers
      additionalProperties: false
      description: ID soal harus unik, berasal dari snapshot attempt, dan option harus milik soal. Maksimal sebanyak
        soal dalam attempt.
      example:
        answers:
        - question_id: 1
          selected_option_id: 11
    SubmitAnswers:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: '#/components/schemas/Answer'
          minItems: 1
      additionalProperties: false
      description: Kirim {} untuk submit jawaban tersimpan. Jika answers dikirim, array tidak boleh kosong. Field
        nilai/identitas ditolak.
      example: {}
...